# backend/artcryption-asset-management/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/assets":{"post":{"operationId":"AssetsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssetDto"}}}},"responses":{"201":{"description":""}},"tags":["Assets"]}},"/assets/users/{userId}":{"get":{"operationId":"AssetsController_findAll","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"collectionId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Get All User Assets","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssetDto"}}}}},"tags":["Assets"]}},"/assets/{id}":{"get":{"operationId":"AssetsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assets"]},"patch":{"operationId":"AssetsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAssetDto"}}}},"responses":{"200":{"description":""}},"tags":["Assets"]},"delete":{"operationId":"AssetsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assets"]}},"/collections":{"post":{"operationId":"CollectionsController_create","summary":"Add a collection","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCollectionDto"}}}},"responses":{"403":{"description":"Forbidden."}},"tags":["Asset Collections"]}},"/collections/users/{userId}":{"get":{"operationId":"CollectionsController_findAll","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get All user Collections","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCollectionDto"}}}}},"tags":["Asset Collections"]}},"/collections/{id}":{"get":{"operationId":"CollectionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get one collection","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCollectionDto"}}}}},"tags":["Asset Collections"]},"patch":{"operationId":"CollectionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCollectionDto"}}}},"responses":{"200":{"description":"Update collection"}},"tags":["Asset Collections"]},"delete":{"operationId":"CollectionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete a collection"}},"tags":["Asset Collections"]}}},"info":{"title":"Artcryption Assets Management  APIs","description":"The Short Artcryption  API description","version":"1.0","contact":{}},"tags":[{"name":"Artcryption","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateAssetDto":{"type":"object","properties":{"externalLink":{"type":"string"},"ownerId":{"type":"string"},"collectionId":{"type":"string"},"createdBy":{"type":"string"}},"required":["externalLink","ownerId","collectionId","createdBy"]},"UpdateAssetDto":{"type":"object","properties":{"title":{"type":"string"},"titleOfWork":{"type":"string"},"traits":{"type":"string"},"placeOfPblication":{"type":"string"},"publicationDate":{"type":"string"},"dimensions":{"type":"string"},"materials":{"type":"string"},"editionNumber":{"type":"string"},"numberOfEditions":{"type":"string"},"tokenAddress":{"type":"string"},"approvedStatusId":{"type":"string"},"isAddedToFeed":{"type":"boolean"},"isBlockchainEnabled":{"type":"boolean"},"isMinted":{"type":"boolean"},"isPrivate":{"type":"boolean"},"AssetType":{"type":"number"},"description":{"type":"string"},"updatedBy":{"type":"string"}},"required":["title","titleOfWork","traits","placeOfPblication","publicationDate","dimensions","materials","editionNumber","numberOfEditions","tokenAddress","approvedStatusId","isAddedToFeed","isBlockchainEnabled","isMinted","isPrivate","AssetType","description","updatedBy"]},"CreateCollectionDto":{"type":"object","properties":{"name":{"type":"string"},"shortDescription":{"type":"string"},"description":{"type":"string"},"private":{"type":"boolean"},"createdBy":{"type":"string"}},"required":["name","shortDescription","description","private","createdBy"]},"UpdateCollectionDto":{"type":"object","properties":{"updatedBy":{"type":"string"}},"required":["updatedBy"]}}}}